Renaming Navar.jsx to Navbar.jsx
