+++ src/App.js
@@ -1,6 +1,6 @@
 // import { Routes, Route } from "react-router-dom";
 import Footer from "./components/footer/Footer";
-import Navbar from "./components/Navbar/Navar";
+import Navbar from "./components/navbar/Navbar";
 import Home from "./pages/Home/Home";
 
 function App() {
